requirements.txt
# Image Processing (Latest versions support Python 3.14 wheels)
pillow>=11.0.0
opencv-python>=4.11.0
pytesseract>=0.3.13

# Core Math & Data
numpy>=2.1.0
SQLAlchemy>=2.0.35
pymysql>=1.1.1

# AI & Reporting
google-generativeai>=0.8.6
reportlab>=4.2.5

# UI & Environment
streamlit>=1.42.0
python-dotenv>=1.0.1
bcrypt>=4.2.0

# Testing
pytest>=8.0.0